Create Simple delegating mechanism
In this example, users can delegate their mining power to others. The delegate_mining_power method allows delegators to specify the delegate's address and the amount of mining power to delegate. The delegate can then mine on behalf of the delegator with the specified mining power. This is a simplified representation and can be expanded upon for more complex scenarios.